Initial commit, quick and dirty solution
Creating a library using ISBNs is operating pretty well as the isbnlib is very robust. The ISSN library is very ugly and poorly refined right now as it uses the crossrefapi which isn't as clear as isbnlib. Needs testing with actual individual journal entries to find out what info is needed to differentiate between entries, as year and many other data points aren't gathered currently.